add: top bar on main
